heading 2 css
Finished the heading 2 css by changing the color and size of the font as well as centering it. I also finished the css for class bolded which made some letters pop out more. Lastly, I changed the weight and size of the font for heading 1 to still make it less noticeable and distract user from form.